fix: `VariantMapper` of `AssemblyMapper._norm` is configurated differently
`AssemblyMapper._norm` is a `Normalizer` object. `Normalizer` has a field `vm` of `VariantMapper` class. In the current code, when initialize an `AssemblyMapper` object with the argument set to a different value from `global_config`, the `vm` will be created using `global_config`. 

The fix passes a `VariantMapper` object initialized using the same arguments as `AssemblyMapper` object.